...G...... [ZBX-9847] fix compilation warning
git-svn-id: svn://svn.zabbix.com/branches/dev@55610 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBX-9847/src/libs/zbxsysinfo/common/cpu.c
@@ -59,7 +59,7 @@ int	SYSTEM_CPU_DISCOVERY(AGENT_REQUEST *request, AGENT_RESULT *result)
 		zbx_json_addobject(&json, NULL);
 
 		zbx_json_adduint64(&json, "{#CPU.NUMBER}", cpus.values[i].first);
-		zbx_json_addstring(&json, "{#CPU.STATUS}", get_cpu_status_string(cpus.values[i].second),
+		zbx_json_addstring(&json, "{#CPU.STATUS}", get_cpu_status_string((int)cpus.values[i].second),
 				ZBX_JSON_TYPE_STRING);
 
 		zbx_json_close(&json);
